Refactor Item and ShoppingBasket to use BigDecimal for unit price and total price calculations; update applicable discount logic and tests

+++ src/main/java/com/ibanfr/shoppingbasket/Item.java
@@ -1,5 +1,7 @@
 package com.ibanfr.shoppingbasket;
 
-public record Item(int unitPrice) {
+import java.math.BigDecimal;
+
+public record Item(BigDecimal unitPrice) {
 
 }
